Move out the getter functions from the getter class
