+++ lib/crypto-price/dataSources/coinMarketCap.ts
@@ -1,0 +1,52 @@
+import { CryptoCurrency, FiatCurrency, IPairQuote } from 'models/Currency';
+import fetch from 'adapters/http/fetch';
+
+const CoinMarketCapCryptoMapping: Record<CryptoCurrency, number> = {
+  AVAX: 5805,
+  BNB: 1839,
+  CELO: 5567,
+  ETH: 1027,
+  FTM: 3513,
+  MATIC: 3890,
+  ONE: 3945,
+  XDAI: 8635
+};
+
+const apiServiceToken = process.env.CMC_API_TOKEN!;
+
+export function getPriceFromCoinMarketCap (
+  base: CryptoCurrency,
+  quote: FiatCurrency,
+  apiToken = apiServiceToken
+): Promise<IPairQuote> {
+  return new Promise((resolve, reject) => {
+    const cmcCryptoCurrencyId = CoinMarketCapCryptoMapping[base];
+    fetch(`https://pro-api.coinmarketcap.com/v2/cryptocurrency/quotes/latest?id=${cmcCryptoCurrencyId}&convert=${quote}`, {
+      headers: { 'X-CMC_PRO_API_KEY': apiToken }
+    })
+      .then(data => {
+
+        console.log('CMC', data);
+
+        const quotedPrice = data.data[cmcCryptoCurrencyId.toString()].quote[quote].price;
+
+        if (!quotedPrice) {
+          reject(new Error('No price found'));
+          return;
+        }
+
+        const pairQuote: IPairQuote = {
+          base,
+          quote,
+          amount: quotedPrice,
+          receivedOn: Date.now()
+        };
+
+        resolve(pairQuote);
+
+      })
+      .catch(error => {
+        reject(error);
+      });
+  });
+}
