+++ lib/crypto-price/dataSources/index.ts
@@ -1,0 +1,2 @@
+export * from './coinMarketCap';
+export * from './cryptoCompare';
